+++ frontend/src/lib/openFullScreenPopup.ts
@@ -1,0 +1,5 @@
+const openFullScreenPopup = (url: string) => {
+    const features = `width=${screen.availWidth},height=${screen.availHeight},top=0,left=0,scrollbars=yes,resizable=yes`;
+    window.open(url, "_blank", features);
+};
+export default openFullScreenPopup;