+++ frontend/src/lib/tech_rs_statuses.ts
@@ -5,9 +5,9 @@ const repairshopr_statuses_techs = [
     { id: 7, _status: "Customer Reply" },
     { id: 8, _status: "Assigned to Tech" },
     { id: 9, _status: "Parts Request 1st Approval" },
-    { id: 10, _status: "Quote Pending" },
+    // { id: 10, _status: "Quote Pending" },
     { id: 11, _status: "QC" },
-    { id: 15, _status: "For Invoicing" },
+    // { id: 15, _status: "For Invoicing" },
     { id: 17, _status: "QC Failed" },
     { id: 22, _status: "Parts to be ordered" },
 ];