+++ server/controllers/department/parts/get_parts.js
@@ -9,7 +9,7 @@ const GetPartsForTask = async (req, res) => {
             "SELECT id, unique_id, ticket_number, part_name, part_desc, seal_number, part_quantity, parts_status, created_at, compensation, created_by, updated_at FROM  parts_for_tasks WHERE task_row_id = $1",
             [id]
         );
-        return res.json(rows);
+        return res.status(200).json(rows);
     } catch (err) {
         return res.status(500).json({ error: "Failed to get parts" });
     }